requirements.txt
Flask==2.2.2
Flask_Bcrypt==1.0.1
Flask_JWT_Extended==4.4.4
flask_sqlalchemy==3.0.3
Flask_WTF==1.0.1
plotly==5.13.0
psycopg2_binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
SQLAlchemy==1.4.44
WTForms==3.0.1
email-validator==1.3.1